+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Please file a bug if you notice a violation of semantic versioning.
 
 ## [Unreleased]
 ### Added
+- .gitlab-ci.yml documentation (in example)
 ### Changed
 ### Deprecated
 ### Removed